server/requirements.txt
boto3
botocore
flask
flask-bcrypt
flask-cors
flask-session
moto[dynamodb]
pytest
pytest-ordering
python-dotenv
redis
requests